fix: catch bug!
