+++ Makefile
@@ -6,7 +6,7 @@
 #    By: wonyang <wonyang@student.42seoul.kr>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/03/04 21:52:18 by wonyang           #+#    #+#              #
-#    Updated: 2023/03/31 17:41:49 by wonyang          ###   ########seoul.kr   #
+#    Updated: 2023/04/04 17:51:13 by wonyang          ###   ########seoul.kr   #
 #                                                                              #
 # **************************************************************************** #
 
@@ -92,7 +92,8 @@ BONUS_SRCS		= $(patsubst %.c, %_bonus.c, $(_BONUS_SRCS)) \
 				  bonus/parsing/door_sprite_bonus.c \
 				  bonus/parsing/divide_lst_type_bonus.c \
 				  bonus/parsing/check_bonus.c \
-				  bonus/render/sprite_bonus.c
+				  bonus/render/sprite_bonus.c \
+				  bonus/render/render_sprite_bonus.c
 				  
 BONUS_HEAD		= -Ibonus \
 				  $(HEADERS)
